Refactor deploy module: divide into different modules
