+++ apiserver/paasng/paasng/engine/deploy/steps/image_release.py
@@ -23,10 +23,11 @@ from django.utils.translation import gettext as _
 
 from paasng.dev_resources.servicehub.manager import mixed_service_mgr
 from paasng.engine.constants import JobStatus
-from paasng.engine.deploy.credentials import ImageCredentialManager
-from paasng.engine.deploy.infras import DeployProcedure, DeployStep, Style
-from paasng.engine.deploy.pre_release import ApplicationPreReleaseExecutor
-from paasng.engine.deploy.preparations import get_app_description_handler, get_processes
+from paasng.engine.deploy.config import ImageCredentialManager
+from paasng.engine.deploy.infra.output import Style
+from paasng.engine.deploy.infra.source import get_app_description_handler, get_processes
+from paasng.engine.deploy.steps.pre_release import ApplicationPreReleaseExecutor
+from paasng.engine.deploy.workflow import DeployProcedure, DeployStep
 from paasng.engine.models import DeployPhaseTypes
 from paasng.engine.signals import post_phase_end, pre_phase_start
 from paasng.extensions.declarative.exceptions import ControllerError, DescriptionValidationError
