+++ apiserver/paasng/paasng/engine/deploy/steps/release.py
@@ -24,9 +24,10 @@ from typing import Optional
 from django.utils.translation import gettext as _
 
 from paasng.engine.constants import JobStatus
+from paasng.engine.deploy.config import AppDefaultDomains, AppDefaultSubpaths, get_env_variables
 from paasng.engine.deploy.engine_svc import EngineDeployClient
-from paasng.engine.deploy.infras import AppDefaultDomains, AppDefaultSubpaths, DeployStep, get_env_variables
-from paasng.engine.deploy.preparations import get_processes_by_build, update_engine_app_config
+from paasng.engine.deploy.infra.models_utils import get_processes_by_build, update_engine_app_config
+from paasng.engine.deploy.workflow import DeployStep
 from paasng.engine.models.deployment import Deployment
 from paasng.engine.models.phases import DeployPhaseTypes
 from paasng.engine.models.processes import ProcessManager
