+++ apiserver/paasng/tests/engine/deploy/infra/test_source.py
@@ -20,61 +20,31 @@ import pathlib
 from typing import Dict, Optional
 from unittest import mock
 
-import gitlab.exceptions
 import pytest
 import yaml
 from django.conf import settings
 from django.test.utils import override_settings
 from django_dynamic_fixture import G
 
-from paasng.accessories.iam.helpers import add_role_members, remove_user_all_roles
-from paasng.accounts.models import Oauth2TokenHolder, UserProfile
 from paasng.dev_resources.sourcectl.exceptions import DoesNotExistsOnServer
-from paasng.dev_resources.sourcectl.models import GitProject, SourcePackage
-from paasng.dev_resources.sourcectl.source_types import get_sourcectl_names
+from paasng.dev_resources.sourcectl.models import SourcePackage
 from paasng.dev_resources.sourcectl.utils import generate_temp_dir, generate_temp_file
-from paasng.engine.constants import DeployConditions
 from paasng.engine.deploy.exceptions import DeployShouldAbortError
-from paasng.engine.deploy.infras import ConsoleStream
-from paasng.engine.deploy.preparations import (
+from paasng.engine.deploy.infra.output import ConsoleStream
+from paasng.engine.deploy.infra.source import (
     check_source_package,
     download_source_to_dir,
     get_app_description_handler,
     get_processes,
-    get_processes_by_build,
     get_source_package_path,
-    update_engine_app_config,
-)
-from paasng.engine.deploy.protections import (
-    EnvProtectionCondition,
-    ModuleEnvDeployInspector,
-    PluginTagValidationCondition,
-    ProductInfoCondition,
-    RepoAccessCondition,
 )
 from paasng.engine.models import Deployment
-from paasng.extensions.bk_plugins.models import BkPluginTag
 from paasng.extensions.declarative.constants import CELERY_BEAT_PROCESS, CELERY_PROCESS, WEB_PROCESS
 from paasng.extensions.declarative.deployment.controller import DeploymentDescription
 from paasng.extensions.declarative.handlers import get_desc_handler
-from paasng.platform.applications.constants import ApplicationRole
-from paasng.platform.core.protections.exceptions import ConditionNotMatched
-from paasng.platform.environments.constants import EnvRoleOperation
-from paasng.platform.environments.models import EnvRoleProtection
 from paasng.platform.modules.constants import SourceOrigin
-from paasng.publish.market.models import Product
 
 pytestmark = pytest.mark.django_db
-
-
-@pytest.fixture()
-def git_client(bk_module):
-    """A fixture used to mock Git repo dependency"""
-    with mock.patch.object(bk_module, "get_source_obj"), mock.patch.object(
-        GitProject, "parse_from_repo_url"
-    ) as get_backends, mock.patch("paasng.dev_resources.sourcectl.controllers.gitlab.GitLabApiClient") as client:
-        get_backends.return_value = GitProject(name="baz", namespace="bar", type='dft_gitlab')
-        yield client()
 
 
 @pytest.fixture(autouse=True)
@@ -270,18 +240,6 @@ class TestGetProcesses:
         assert processes == {"web": "gunicorn wsgi -w 4", "worker": "celery"}
 
 
-def test_get_processes_by_build(bk_module):
-    engine_app = bk_module.envs.get(environment='prod').engine_app
-    fake_procfile = {
-        'web': (
-            'gunicorn wsgi -w 4 -b :$PORT --access-logfile - --error-logfile - --access-logformat '
-            '\'[%(h)s] %({request_id}i)s %(u)s %(t)s "%(r)s" %(s)s %(D)s %(b)s "%(f)s" "%(a)s"\''
-        )
-    }
-    with mock.patch("paasng.engine.deploy.engine_svc.EngineDeployClient.get_procfile", return_value=fake_procfile):
-        assert 'web' in get_processes_by_build(engine_app=engine_app, build_id="fake-build-id")
-
-
 class TestGetSourcePackagePath:
     """Test get_source_package_path()"""
 
@@ -327,8 +285,8 @@ class TestDownloadSourceToDir:
 
     @pytest.fixture(autouse=True)
     def mocked_ctl(self):
-        with mock.patch('paasng.engine.deploy.preparations.get_repo_controller'), mock.patch(
-            'paasng.engine.deploy.preparations.PackageController'
+        with mock.patch('paasng.engine.deploy.infra.source.get_repo_controller'), mock.patch(
+            'paasng.engine.deploy.infra.source.PackageController'
         ):
             yield
 
@@ -417,196 +375,3 @@ class TestCheckSourcePackage:
 
             out, err = capsys.readouterr()
             assert out
-
-
-class TestUpdateEngineAppConfig:
-    """Test update_engine_app_config()"""
-
-    def test_normal(self, bk_module, bk_deployment):
-        with mock.patch('paasng.engine.deploy.preparations.EngineDeployClient') as mocked_client:
-            update_engine_app_config(bk_module.get_envs("prod").engine_app, bk_deployment.version_info)
-            assert mocked_client().update_config.called
-
-
-class TestProductInfoCondition:
-    @pytest.mark.parametrize(
-        "env, create_product, ok",
-        [("prod", False, False), ("prod", True, True), ("stag", False, True), ("stag", True, True)],
-    )
-    def test_validate(self, bk_user, bk_module, env, create_product, ok):
-        application = bk_module.application
-        env = bk_module.get_envs(env)
-
-        if create_product:
-            G(Product, application=application)
-
-        if ok:
-            ProductInfoCondition(bk_user, env).validate()
-        else:
-            with pytest.raises(ConditionNotMatched) as exc_info:
-                ProductInfoCondition(bk_user, env).validate()
-
-            assert exc_info.value.action_name == DeployConditions.FILL_PRODUCT_INFO.value
-
-
-class TestPluginTagCondition:
-    @pytest.mark.parametrize(
-        "env, create_tag, ok",
-        [("prod", False, False), ('prod', True, True), ("stag", False, True), ('stag', True, True)],
-    )
-    def test_validate(self, bk_user, bk_plugin_app, env, create_tag, ok):
-        bk_plugin_profile = bk_plugin_app.bk_plugin_profile
-
-        if create_tag:
-            tag_1 = G(BkPluginTag, code_name='sample-tag-code-1', name='sample-tag-1')
-            bk_plugin_profile.tag = tag_1
-
-        env = bk_plugin_app.get_default_module().envs.get(environment=env)
-        if ok:
-            PluginTagValidationCondition(bk_user, env).validate
-        else:
-            with pytest.raises(ConditionNotMatched) as exc_info:
-                PluginTagValidationCondition(bk_user, env).validate()
-
-            assert exc_info.value.action_name == DeployConditions.FILL_PLUGIN_TAG_INFO.value
-
-
-class TestEnvProtectionCondition:
-    @pytest.mark.parametrize(
-        "user_role, allowed_roles, ok",
-        [
-            (ApplicationRole.ADMINISTRATOR, [], True),
-            (ApplicationRole.ADMINISTRATOR, [ApplicationRole.ADMINISTRATOR], True),
-            (ApplicationRole.ADMINISTRATOR, [ApplicationRole.DEVELOPER], False),
-            (ApplicationRole.DEVELOPER, [], True),
-            (ApplicationRole.DEVELOPER, [ApplicationRole.ADMINISTRATOR], False),
-            (ApplicationRole.DEVELOPER, [ApplicationRole.DEVELOPER], True),
-            (ApplicationRole.DEVELOPER, [ApplicationRole.ADMINISTRATOR, ApplicationRole.DEVELOPER], True),
-        ],
-    )
-    def test_validate(self, bk_user, bk_module, user_role, allowed_roles, ok):
-        application = bk_module.application
-        env = bk_module.get_envs("stag")
-
-        remove_user_all_roles(application.code, bk_user.username)
-        add_role_members(application.code, user_role, bk_user.username)
-
-        for role in allowed_roles:
-            EnvRoleProtection.objects.create(
-                allowed_role=role.value, module_env=env, operation=EnvRoleOperation.DEPLOY.value
-            )
-
-        if ok:
-            EnvProtectionCondition(bk_user, env).validate()
-        else:
-            with pytest.raises(ConditionNotMatched):
-                EnvProtectionCondition(bk_user, env).validate()
-            inspector = ModuleEnvDeployInspector(bk_user, env)
-            inspector.conditions = [EnvProtectionCondition(bk_user, env)]
-            assert [item.action_name for item in inspector.perform().failed_conditions] == [
-                EnvProtectionCondition.action_name
-            ]
-
-
-class TestRepoAccessCondition:
-    @pytest.mark.parametrize(
-        "source_type, create_user_profile, create_token_holder, ok, action_name",
-        [
-            ('dft_gitlab', False, False, False, DeployConditions.NEED_TO_BIND_OAUTH_INFO),
-            ('dft_gitlab', True, False, False, DeployConditions.NEED_TO_BIND_OAUTH_INFO),
-            ('dft_gitlab', True, True, False, DeployConditions.DONT_HAVE_ENOUGH_PERMISSIONS),
-            ('dft_gitlab', True, True, True, ...),
-        ],
-    )
-    def test_validate(
-        self, bk_user, bk_module, git_client, source_type, create_user_profile, create_token_holder, ok, action_name
-    ):
-        env = bk_module.get_envs("stag")
-        bk_module.source_type = source_type
-        bk_module.save()
-        if create_user_profile:
-            profile = G(UserProfile, user=bk_user)
-            if create_token_holder:
-                G(Oauth2TokenHolder, user=profile, provider=get_sourcectl_names().GitLab)
-
-        if ok:
-            assert RepoAccessCondition(bk_user, env).validate() is None
-        else:
-
-            def get_project_info(project):
-                raise gitlab.exceptions.GitlabAuthenticationError
-
-            git_client.get_project_info.side_effect = get_project_info
-            with pytest.raises(ConditionNotMatched) as exc_info:
-                RepoAccessCondition(bk_user, env).validate()
-
-            assert exc_info.value.action_name == action_name.value
-            inspector = ModuleEnvDeployInspector(bk_user, env)
-            inspector.conditions = [RepoAccessCondition(bk_user, env)]
-            assert [item.action_name for item in inspector.perform().failed_conditions] == [action_name.value]
-
-
-class TestModuleEnvDeployInspector:
-    @pytest.mark.parametrize(
-        "user_role, allowed_roles, create_token, create_product, expected",
-        [
-            (
-                ApplicationRole.DEVELOPER,
-                [ApplicationRole.ADMINISTRATOR],
-                False,
-                False,
-                [
-                    DeployConditions.FILL_PRODUCT_INFO,
-                    DeployConditions.CHECK_ENV_PROTECTION,
-                    DeployConditions.NEED_TO_BIND_OAUTH_INFO,
-                ],
-            ),
-            (
-                ApplicationRole.OPERATOR,
-                [ApplicationRole.ADMINISTRATOR],
-                True,
-                False,
-                [DeployConditions.FILL_PRODUCT_INFO, DeployConditions.CHECK_ENV_PROTECTION],
-            ),
-            (
-                ApplicationRole.ADMINISTRATOR,
-                ...,
-                True,
-                False,
-                [DeployConditions.FILL_PRODUCT_INFO],
-            ),
-            (
-                ...,
-                ...,
-                True,
-                True,
-                [],
-            ),
-        ],
-    )
-    def test(self, bk_user, bk_module, git_client, user_role, allowed_roles, create_token, create_product, expected):
-        application = bk_module.application
-        env = bk_module.get_envs("prod")
-        bk_module.source_type = get_sourcectl_names().GitLab
-        bk_module.save()
-
-        if user_role is not ...:
-            remove_user_all_roles(application.code, bk_user.username)
-            add_role_members(application.code, user_role, bk_user.username)
-
-        if allowed_roles is not ...:
-            for role in allowed_roles:
-                EnvRoleProtection.objects.create(
-                    allowed_role=role.value, module_env=env, operation=EnvRoleOperation.DEPLOY.value
-                )
-
-        if create_token:
-            profile = G(UserProfile, user=bk_user)
-            G(Oauth2TokenHolder, user=profile, provider=get_sourcectl_names().GitLab)
-
-        if create_product:
-            G(Product, application=application)
-
-        inspector = ModuleEnvDeployInspector(bk_user, env)
-        assert [item.action_name for item in inspector.perform().failed_conditions] == [c.value for c in expected]
-        assert inspector.all_matched is not len(expected)
